+++ astronomer/providers/google/cloud/hooks/dataproc.py
@@ -1,35 +1,115 @@
 import warnings
+from abc import ABC
 from typing import Any, Optional, Sequence, Tuple, Union
 
-from airflow.providers.google.cloud.hooks.dataproc import DataprocHook
+from airflow.providers.google.common.consts import CLIENT_INFO
 from airflow.providers.google.common.hooks.base_google import GoogleBaseHook
-from asgiref.sync import sync_to_async
+from google.api_core import gapic_v1
+from google.api_core.client_options import ClientOptions
 from google.api_core.retry import Retry
-from google.cloud.dataproc_v1 import Job
+from google.cloud.dataproc_v1 import (
+    ClusterControllerAsyncClient,
+    Job,
+    JobControllerAsyncClient,
+)
+from google.cloud.dataproc_v1.types import clusters
 
 JobType = Union[Job, Any]
 
 
-class DataprocHookAsync(DataprocHook):
-    """
-    Async Hook for Google Cloud Dataproc APIs.
-    All the methods in the hook where project_id is used must be called with
-    keyword arguments rather than positional.
-    """
+class DataprocHookAsync(GoogleBaseHook, ABC):
+    """Async Hook for Google Cloud Dataproc APIs"""
+
+    def get_cluster_client(
+        self, region: Optional[str] = None, location: Optional[str] = None
+    ) -> ClusterControllerAsyncClient:
+        """
+        Get async cluster controller client for GCP Dataproc.
+
+        :param region: The Cloud Dataproc region in which to handle the request.
+        :param location: (To be deprecated). The Cloud Dataproc region in which to handle the request.
+        """
+        if location is not None:
+            warnings.warn(
+                "Parameter `location` will be deprecated. "
+                "Please provide value through `region` parameter instead.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+            region = location
+        client_options = ClientOptions()
+        if region and region != "global":
+            client_options = ClientOptions(api_endpoint=f"{region}-dataproc.googleapis.com:443")
+
+        return ClusterControllerAsyncClient(
+            credentials=self._get_credentials(), client_info=CLIENT_INFO, client_options=client_options
+        )
+
+    def get_job_client(
+        self, region: Optional[str] = None, location: Optional[str] = None
+    ) -> JobControllerAsyncClient:
+        """
+        Get async job controller for GCP Dataproc.
+
+        :param region: The Cloud Dataproc region in which to handle the request.
+        :param location: (To be deprecated). The Cloud Dataproc region in which to handle the request.
+        """
+        if location is not None:
+            warnings.warn(
+                "Parameter `location` will be deprecated. "
+                "Please provide value through `region` parameter instead.",
+                DeprecationWarning,
+                stacklevel=2,
+            )
+            region = location
+        client_options = ClientOptions()
+        if region and region != "global":
+            client_options = ClientOptions(api_endpoint=f"{region}-dataproc.googleapis.com:443")
+
+        return JobControllerAsyncClient(
+            credentials=self._get_credentials(), client_info=CLIENT_INFO, client_options=client_options
+        )
+
+    async def get_cluster(
+        self,
+        region: str,
+        cluster_name: str,
+        project_id: str,
+        retry: Union[Retry, gapic_v1.method._MethodDefault] = gapic_v1.method.DEFAULT,
+        metadata: Sequence[Tuple[str, str]] = (),
+    ) -> clusters.Cluster:
+        """
+        Get a cluster details from GCP using `ClusterControllerAsyncClient`
+
+        :param region: The Cloud Dataproc region in which to handle the request
+        :param cluster_name: The name of the cluster
+        :param project_id:  The ID of the Google Cloud project the cluster belongs to
+        :param retry: A retry object used to retry requests. If ``None`` is specified, requests will not be
+            retried
+        :param metadata: Additional metadata that is provided to the method
+        """
+        client = self.get_cluster_client(region=region)
+        return await client.get_cluster(
+            region=region,
+            cluster_name=cluster_name,
+            project_id=project_id,
+            retry=retry,
+            metadata=metadata,
+        )
 
     @GoogleBaseHook.fallback_to_default_project_id
-    async def get_job(  # type: ignore[override]
+    async def get_job(
         self,
         job_id: str,
         project_id: str,
+        timeout: float = 5,
         region: Optional[str] = None,
         location: Optional[str] = None,
-        retry: Optional[Retry] = None,
-        timeout: Optional[float] = None,
+        retry: Union[Retry, gapic_v1.method._MethodDefault] = gapic_v1.method.DEFAULT,
         metadata: Sequence[Tuple[str, str]] = (),
     ) -> JobType:
         """
-        Gets the resource representation for a job in a project.
+        Gets the resource representation for a job using `JobControllerAsyncClient`.
 
         :param job_id: Id of the Dataproc job
         :param project_id: Required. The ID of the Google Cloud project the cluster belongs to.
@@ -52,8 +132,8 @@ class DataprocHookAsync(DataprocHook):
                 region = location
             else:
                 raise TypeError("missing 1 required keyword argument: 'region'")
-        client = await sync_to_async(self.get_job_client)(region=region)
-        job = client.get_job(
+        client = self.get_job_client(region=region)
+        job = await client.get_job(
             request={"project_id": project_id, "region": region, "job_id": job_id},
             retry=retry,
             timeout=timeout,
